New and improved refactor
