+++ ci/snapshot/image_managers/snapshot-canary
